msm: move test helper structs into helpers_test.go
Addresses review comment: msm_test.go was cluttered with the fakeVMBlock,
blockStore, signatureVerifier/aggregator, validatorSetRetriever,
keyAggregator, blockBuilder stubs plus newStateMachine/testConfig. They
are test plumbing, not test cases, so relocate them to
helpers_test.go to make the actual tests easier to navigate.
